requirements.txt
streamlit
llama-index-core
llama-index-llms-gemini
llama-index-embeddings-gemini
python-dotenv
pypdf